requirements-dev.txt
asgiref==3.2.10
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-static==0.0.6
dj-database-url
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==4.2.0
openapi-codec==1.3.2
packaging==20.4
pluggy==0.6.0
py==1.5.3
PyJWT==1.7.1
pyparsing==2.4.7
pytest==3.6.2
python-decouple==3.3
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.11.0
sqlparse==0.3.1
static3==0.7.0
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.1.0
